requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
gunicorn==22.0.0
httpx==0.27.2
Pillow==10.4.0
opencv-contrib-python-headless==4.10.0.84
pyzbar==0.1.9
